refactoring
Changes to be committed:
	new file:   Pipfile
	new file:   Pipfile.lock
	modified:   README.md
	modified:   app.py
	deleted:    create_simple_DICOMDIR.py
	new file:   folder_reader.py
