+++ Pipfile
@@ -1,0 +1,15 @@
+[[source]]
+name = "pypi"
+url = "https://pypi.org/simple"
+verify_ssl = true
+
+[dev-packages]
+
+[packages]
+python-box = "==4.0.4"
+pydicom = "==1.4.1"
+Pillow = "==7.0.0"
+PySimpleGUI = "==4.15.2"
+
+[requires]
+#python_version = "3.8"
