+++ README.md
@@ -1,46 +1,56 @@
 # CT DICOM Images
 
-1. classify a folder of DICOM images(even with subfolders) by 4-level hierachy `patient/study/series/instance`.  
+1. select a folder containing DICOM images
 
-   It's inspired from [The Structure of DICOMDIR](https://www.medicalconnections.co.uk/kb/DICOMDIR/) and [Read DICOM directory](https://pydicom.github.io/pydicom/stable/auto_examples/input_output/plot_read_dicom_directory.html#sphx-glr-auto-examples-input-output-plot-read-dicom-directory-py).
+   We walk the folder and organize DICOM images by a 4-level hierarchy `patient/study/series/instance`.
 
-   **Note:** In pydicom V1.4, it does not support to create DICOMDIR file and the feature will be introduced in future 2.0 version. The example of reading DICOMDIR uses the file which is generated by `dcmtk`. 
+   It's inspired from [The Structure of DICOMDIR](https://www.medicalconnections.co.uk/kb/DICOMDIR/) and [Read DICOM directory](https://pydicom.github.io/pydicom/stable/auto_examples/input_output/plot_read_dicom_directory.html).
 
-2. organize the classified DICOM images as tree view.
+   `DicomFolder` is should have a similar API as `DicomDir`.
 
-   use PySimpleGUI - tkinter to show tree view.
+   **Note:** In pydicom 1.4, it does not support creating DICOMDIR and the feature will be introduced in future version (2.0). The example of reading DICOMDIR uses the DICOMDIR generated by `dcmtk`.
 
+2. render `DicomFolder` as tree view
 
+   This version is using PySimpleGUI - tkinter to show tree view.
 
-PySimpleGUI - tkinter version:
-[PySimpleGUI](https://github.com/PySimpleGUI/PySimpleGUI)
-[Tree Element](https://github.com/PySimpleGUI/PySimpleGUI#tree-element)
+   PySimpleGUI - tkinter version:  
+   [PySimpleGUI](https://github.com/PySimpleGUI/PySimpleGUI)  
+   [Tree Element](https://github.com/PySimpleGUI/PySimpleGUI#tree-element)
 
-QT QML Treeview:
-[Tutorial Qt Creator - QML - TreeView](https://www.youtube.com/watch?v=J_jMDro3sMg)
-[empty TreeView in QML from jsonModel](https://github.com/eyllanesc/stackoverflow/tree/master/questions/50007170)
-[TreeModelFromJSONApp](https://gitlab.com/eska2000/TreeModelFromJSONApp)
-[QML TreeView](https://ruedigergad.com/2011/08/14/qml-treeview/)
-[Create Model for QML TreeView](https://stackoverflow.com/questions/45166367/create-model-for-qml-treeview)
-[Qt qml treeview 树控件](https://www.cnblogs.com/surfsky/p/4309299.html)
-[QML自定义树控件（TreeView 的style加上节点可拖动）](https://blog.csdn.net/weixin_40912639/article/details/83962250)
-QT QWidget Treeview:
-[PyQt5 Treeview](https://pythonspot.com/pyqt5-treeview/)
+   QT QML Treeview:  
+   [Tutorial Qt Creator - QML - TreeView](https://www.youtube.com/watch?v=J_jMDro3sMg)  
+   [empty TreeView in QML from jsonModel](https://github.com/eyllanesc/stackoverflow/tree/master/questions/50007170)  
+   [TreeModelFromJSONApp](https://gitlab.com/eska2000/TreeModelFromJSONApp)  
+   [QML TreeView](https://ruedigergad.com/2011/08/14/qml-treeview/)  
+   [Create Model for QML TreeView](https://stackoverflow.com/questions/45166367/create-model-for-qml-treeview)  
+   [Qt qml treeview 树控件](https://www.cnblogs.com/surfsky/p/4309299.html)  
+   [QML 自定义树控件（TreeView 的 style 加上节点可拖动）](https://blog.csdn.net/weixin_40912639/article/details/83962250)  
+   QT QWidget Treeview:  
+   [PyQt5 Treeview](https://pythonspot.com/pyqt5-treeview/)
+
+3. select a Series and enter `Viewer`
+
+   - first run the algorithm at background worker and get the analysis
+   - have standard DICOM viewer features
+   - overlay the analysis to instance, render ROI in timeline
 
 ## Usage
 
-Requirement: Python3.6+
+Requirement: Python 3.6+, `pipenv`
 
 ```sh
-pip install -r requirements.txt
-python app.py
+pipenv run ./app.py
+# OR
+pipenv shell
+./app.py
 ```
 
 ## DICOM CT Images Sorting
 
 How the DICOM CT images should be sorted is ultimately dependent on the usage context, but as a rule of thumb I would recommend that you first group the images based on (patient), study and series using these tags:
 
-```sh
+```
 (0010,0020)  Patient ID
 (0020,000D)  Study Instance UID
 (0020,000E)  Series Instance UID
@@ -52,3 +62,16 @@ Another alternative is to use the Image Position (Patient) (0020,0032), which is
 
 If the series also contains a localizer image, this image would have to be excluded from positional sorting.
 
+## TODO
+
+- `DicomFolder` only have children while `DicomDir` has `patient_records`
+- Do not show instances in tree  
+  see comments in `Run` event  
+  our business logic should be seperated from UI data structure  
+  (meaning that our own data structure should support our business logic's need)
+- trigger "Scan" after (a different) folder is selected  
+  no need to use "Scan" button  
+  with `enable_events=True`?  
+  idealy selecting the same folder should not trigger "Scan"
+- double click on tree to trigger "View"  
+  this is better than right click menu, which is quite buggy and non-standard
